refactor: centralize auth state in context provider
